requirements.txt
pyspark>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
pymongo>=4.5.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
findspark>=2.0.0
python-dateutil>=2.8.0